academy-chatbot-main/requirements.txt
langgraph
langchain-core
langchain-google-community[gmail]
python-dotenv
ipython
pymupdf
sentence_transformers
python-docx
faiss-cpu
langgraph-checkpoint-sqlite
tiktoken
langchain-openai
langchain-community
langchain-text-splitters
pyyaml
beautifulsoup4
Flask-Cors
Flask
gunicorn
email-validator
bleach
google-cloud-storage
APScheduler
boto3==1.35.95
langchain_experimental